Released versoin 3.4.1 based on 3.3.712 .
git-svn-id: svn://svn.bearnip.com/ldmud/3.4/trunk@2287 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/sprintf.c
@@ -1454,7 +1454,7 @@ static char buff[BUFF_SIZE];         /* For error messages */
     char          format_char;     /* format type */
     unsigned int  nelemno;         /* next offset into array */
     unsigned int  fpos;            /* position in format_str */
-    unsigned int  fs;              /* field size */
+    unsigned long fs;              /* field size */
     int           pres;            /* precision */
     unsigned int  err_num;         /* error code */
     char         *pad;             /* fs pad string */
